examples/registry : add all type of data for vk (except large binary) dunicodestring : add conversion method using icu dvalue : add implicit conversion via operator() to avoid to .get<T>() every time, take care could be ambigous in some case python/dvalue : return DUnicodeString as unicode object, return DBuffer as ByteArray rather than buffer type because DBuffer could delete it's content before it's freed by python python/examples/qregistry : handle display of data value,  ex: hexlify if ByteArray, concatenate string for REG_MULTI_SZ (handled as DVectorString) examples/rpc/serializerpc : handle serialization & deserialization of DBuffer type

+++ dobject.cpp
@@ -57,11 +57,22 @@ void DObject::setValue(DUnicodeString const& name, DValue const& v)
   throw DException(this->instanceOf()->name() + "::setValue instance has no attribute " + name);
 }
 
+//template <typename RealType>
+//DValue DObject::call(DUnicodeString const& name, RealType v)
+//{
+  //int32_t index = this->instanceOf()->findAttribute(name);
+  //DValue value = RealValue<RealType>(v);
+  //if (index != -1)
+   //return (this->call((size_t)index, value));
+
+  //throw DException(this->instanceOf()->name() + "::call instance has no attribute " + name); 
+//}
+
 DValue DObject::call(DUnicodeString const& name, DValue const& v) //const ? XXX  : throw
 {
   int32_t index = this->instanceOf()->findAttribute(name);
   if (index != -1)
-   return (this->call(index, v));
+   return (this->call((size_t)index, v));
 
   throw DException(this->instanceOf()->name() + "::call instance has no attribute " + name); 
 }
@@ -70,7 +81,7 @@ DValue DObject::call(DUnicodeString const& name) //const ? XXX  : throw
 {
   int32_t index = this->instanceOf()->findAttribute(name);
   if (index != -1)
-    return (this->call(index, RealValue<DObject*>(DNone)));
+    return (this->call((size_t)index, RealValue<DObject*>(DNone)));
 
   throw DException(this->instanceOf()->name() + "::call instance has no attribute " + name); 
 }
